backend-server/ml-service/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
requests
pydantic
pymongo
librosa
numpy
python-multipart